+++ crates/cli/src/main.rs
@@ -1,3 +1,27 @@
+// Copyright 2024 mik
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use std::env;
+
+use opla_core::gguf::GGUF;
+
 fn main() {
-    println!("Opla CLI TODO!");
+        println!("Opla CLI WIP");
+    let args: Vec<_> = env::args().collect();
+    if args.len() > 1 {
+        println!("The first argument is {}", args[1]);
+    }
+    let mut gguf = GGUF::new();
+    gguf.read(&args[1]).unwrap();
 }